Adopt LaunchRequestHandler & RestartRequestHandler
