+++ lldb/tools/lldb-dap/Handler/ContinueRequestHandler.cpp
@@ -1,4 +1,4 @@
-//===-- CompletionsHandler.cpp --------------------------------------------===//
+//===-- ContinueRequestHandler.cpp ----------------------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
